Release 2023-12-11

+++ contracts/vortex/CarbonVortex.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: BUSL-1.1
+// SPDX-License-Identifier: SEE LICENSE IN LICENSE
 pragma solidity 0.8.19;
 
 import { ReentrancyGuardUpgradeable } from "@openzeppelin/contracts-upgradeable/security/ReentrancyGuardUpgradeable.sol";
@@ -245,7 +245,7 @@ contract CarbonVortex is ICarbonVortex, Upgradeable, ReentrancyGuardUpgradeable,
     }
 
     function _validateTokens(Token[] calldata tokens) private view {
-        uint len = tokens.length;
+        uint256 len = tokens.length;
         if (len == 0) {
             revert InvalidTokenLength();
         }
